harjoitustyo/pyproject.toml
[tool.poetry]
name = "harjoitustyo"
version = "0.1.0"
description = ""
authors = ["pietarni <pietarni@ad.helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.7.3"
pillow = "^9.3.0"
scikit-image = "^0.19.3"
matplotlib = "^3.6.2"
numpy = "^1.23.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
